requirements.txt
openai
streamlit
pinecone-client
fuzzywuzzy